+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 * `mxprobe` now distinguishes between soft and hard fails
 * `mxsniff` command line now produces valid CSV or JSON (verbose mode)
+* Verbose output now includes canonical representation of an email address to facilitate comparison
 
 0.3.2
 =====